+++ sdk/core/azure-core/src/main/java/com/azure/core/util/ProgressReceiver.java
@@ -4,11 +4,11 @@
 package com.azure.core.util;
 
 /**
- * A {@link ProgressListener} is an interface that can be used to report progress on network transfers.
- * The {@link #onProgress(long)} method will be called periodically with the total number of bytes
+ * A {@link ProgressReceiver} is an interface that can be used to report progress on network transfers.
+ * The {@link #reportProgress(long)} method will be called periodically with the total number of bytes
  * transferred.
  */
-public interface ProgressListener {
+public interface ProgressReceiver {
     /**
      * The callback function invoked as progress is reported.
      *
@@ -19,5 +19,5 @@ public interface ProgressListener {
      *
      * @param bytesTransferred The total number of bytes transferred during this transaction.
      */
-    void onProgress(long bytesTransferred);
+    void reportProgress(long bytesTransferred);
 }