+++ sdk/core/azure-core/CHANGELOG.md
@@ -10,7 +10,7 @@
   - Added `HttpPipelineSyncPolicy` to represent synchronous `HttpPipelinePolicy`.
   - Added `HttpPipelineNextSyncPolicy` to invoke the next synchronous policy in pipeline. to process synchronous policy pipeline.
   - Added `HttpPipelineCallState` to maintain request specific pipeline and contextual data.
-- Added `ProgressReporter` and `ProgressListener` to provide ability to track transfers.
+- Added `ProgressReporter` and `ProgressReceiver` to provide ability to track transfers.
 - Added `Contexts` utility to manipulate known cross-cutting key-value pairs.
   - Added ability to get and set `ProgressReporter` on `Context`.